Create cli for building input baskets
As part of the process of refactoring/extracting some of the metadata
model/types out of adams apple and into a separate library, we still
need a command line interface separate from the app, an separate from
the library for running on the "server" (or developers machines) to
split up the files coming out of our R scripts. This package will
maintain that CLI and any future command line scripts necessary for
building feedme jefferson inputs.