Add support for tracking remaining expansions
Given a core basket, there was no way to know which branches had
available branch expansions. Now we've added ids to the expansions
themselves and a tree of branches that have available expansions along
with their ids. For now the ids are just the branch number, but we can
replace this with some sort of a hash in the future.