Add github npm package repo and get ready for publishing
Switching over from npm linking the feedme-trees dependency to pulling
it from the github npm package repo. Eventually we'll want to publish
this package too for easier use as a locally installed CLI, but I don't
want to do that until we nail down some of the command line arguments --
at the moment I've just been hacking the hardcoded values for appname,
eagerness and frequency.